Julia docstrings

Julia function docstrings (from source code).

StravaAPI.ActivityStats

StravaAPI.ActivityTotal

StravaAPI.ActivityZone

StravaAPI.AltitudeStream

StravaAPI.BaseStream

StravaAPI.CadenceStream

StravaAPI.ClubActivity

StravaAPI.ClubAthlete

StravaAPI.Comment

StravaAPI.DetailedActivity

StravaAPI.DetailedAthlete

StravaAPI.DetailedClub

StravaAPI.DetailedGear

StravaAPI.DetailedSegment

StravaAPI.DetailedSegmentEffort

StravaAPI.DistanceStream

StravaAPI.Error

StravaAPI.ExplorerResponse

StravaAPI.ExplorerSegment

StravaAPI.Fault

StravaAPI.HeartRateZoneRanges

StravaAPI.HeartrateStream

StravaAPI.Lap

StravaAPI.LatLngStream

StravaAPI.MetaActivity

StravaAPI.MetaAthlete

StravaAPI.MetaClub

StravaAPI.MovingStream

StravaAPI.PhotosSummary

StravaAPI.PhotosSummaryPrimary

StravaAPI.PolylineMap

StravaAPI.PowerStream

StravaAPI.PowerZoneRanges

StravaAPI.Route

StravaAPI.SmoothGradeStream

StravaAPI.SmoothVelocityStream

StravaAPI.Split

StravaAPI.StreamSet

StravaAPI.SummaryActivity

StravaAPI.SummaryAthlete

StravaAPI.SummaryClub

StravaAPI.SummaryGear

StravaAPI.SummaryPRSegmentEffort

StravaAPI.SummarySegment

StravaAPI.SummarySegmentEffort

StravaAPI.TemperatureStream

StravaAPI.TimeStream

StravaAPI.TimedZoneRange

StravaAPI.UpdatableActivity

StravaAPI.Upload

StravaAPI.Waypoint

StravaAPI.ZoneRange

StravaAPI.Zones

StravaAPI.basepath

StravaAPI.basepath

StravaAPI.basepath

StravaAPI.basepath

StravaAPI.basepath

StravaAPI.basepath

StravaAPI.basepath

StravaAPI.basepath

StravaAPI.basepath

StravaAPI.create_activity

StravaAPI.create_upload

StravaAPI.explore_segments

StravaAPI.get_activity_by_id

StravaAPI.get_activity_streams

StravaAPI.get_club_activities_by_id

StravaAPI.get_club_admins_by_id

StravaAPI.get_club_by_id

StravaAPI.get_club_members_by_id

StravaAPI.get_comments_by_activity_id

StravaAPI.get_efforts_by_segment_id

StravaAPI.get_gear_by_id

StravaAPI.get_kudoers_by_activity_id

StravaAPI.get_laps_by_activity_id

StravaAPI.get_logged_in_athlete

StravaAPI.get_logged_in_athlete_activities

StravaAPI.get_logged_in_athlete_clubs

StravaAPI.get_logged_in_athlete_starred_segments

StravaAPI.get_logged_in_athlete_zones

StravaAPI.get_route_as_g_p_x

StravaAPI.get_route_as_t_c_x

StravaAPI.get_route_by_id

StravaAPI.get_route_streams

StravaAPI.get_routes_by_athlete_id

StravaAPI.get_segment_by_id

StravaAPI.get_segment_effort_by_id

StravaAPI.get_segment_effort_streams

StravaAPI.get_segment_streams

StravaAPI.get_stats

StravaAPI.get_upload_by_id

StravaAPI.get_zones_by_activity_id

StravaAPI.star_segment

StravaAPI.update_activity_by_id

StravaAPI.update_logged_in_athlete

StravaAPI.ActivityStats — Type

ActivityStats A set of rolled-up statistics and totals for an athlete

ActivityStats(;
    biggest_ride_distance=nothing,
    biggest_climb_elevation_gain=nothing,
    recent_ride_totals=nothing,
    recent_run_totals=nothing,
    recent_swim_totals=nothing,
    ytd_ride_totals=nothing,
    ytd_run_totals=nothing,
    ytd_swim_totals=nothing,
    all_ride_totals=nothing,
    all_run_totals=nothing,
    all_swim_totals=nothing,
)

- biggest_ride_distance::Float64 : The longest distance ridden by the athlete.
- biggest_climb_elevation_gain::Float64 : The highest climb ridden by the athlete.
- recent_ride_totals::ActivityTotal
- recent_run_totals::ActivityTotal
- recent_swim_totals::ActivityTotal
- ytd_ride_totals::ActivityTotal
- ytd_run_totals::ActivityTotal
- ytd_swim_totals::ActivityTotal
- all_ride_totals::ActivityTotal
- all_run_totals::ActivityTotal
- all_swim_totals::ActivityTotal

StravaAPI.ActivityTotal — Type

ActivityTotal A roll-up of metrics pertaining to a set of activities. Values are in seconds and meters.

ActivityTotal(;
    count=nothing,
    distance=nothing,
    moving_time=nothing,
    elapsed_time=nothing,
    elevation_gain=nothing,
    achievement_count=nothing,
)

- count::Int64 : The number of activities considered in this total.
- distance::Float32 : The total distance covered by the considered activities.
- moving_time::Int64 : The total moving time of the considered activities.
- elapsed_time::Int64 : The total elapsed time of the considered activities.
- elevation_gain::Float32 : The total elevation gain of the considered activities.
- achievement_count::Int64 : The total number of achievements of the considered activities.

StravaAPI.ActivityZone — Type

ActivityZone

ActivityZone(;
    score=nothing,
    distribution_buckets=nothing,
    type=nothing,
    sensor_based=nothing,
    points=nothing,
    custom_zones=nothing,
    max=nothing,
)

- score::Int64
- distribution_buckets::Vector{TimedZoneRange} : Stores the exclusive ranges representing zones and the time spent in each.
- type::String
- sensor_based::Bool
- points::Int64
- custom_zones::Bool
- max::Int64

StravaAPI.AltitudeStream — Type

AltitudeStream

AltitudeStream(;
    original_size=nothing,
    resolution=nothing,
    series_type=nothing,
    data=nothing,
)

- original_size::Int64 : The number of data points in this stream
- resolution::String : The level of detail (sampling) in which this stream was returned
- series_type::String : The base series used in the case the stream was downsampled
- data::Vector{Float32} : The sequence of altitude values for this stream, in meters

StravaAPI.BaseStream — Type

BaseStream

BaseStream(;
    original_size=nothing,
    resolution=nothing,
    series_type=nothing,
)

- original_size::Int64 : The number of data points in this stream
- resolution::String : The level of detail (sampling) in which this stream was returned
- series_type::String : The base series used in the case the stream was downsampled

StravaAPI.CadenceStream — Type

CadenceStream

CadenceStream(;
    original_size=nothing,
    resolution=nothing,
    series_type=nothing,
    data=nothing,
)

- original_size::Int64 : The number of data points in this stream
- resolution::String : The level of detail (sampling) in which this stream was returned
- series_type::String : The base series used in the case the stream was downsampled
- data::Vector{Int64} : The sequence of cadence values for this stream, in rotations per minute

StravaAPI.ClubActivity — Type

ClubActivity

ClubActivity(;
    athlete=nothing,
    name=nothing,
    distance=nothing,
    moving_time=nothing,
    elapsed_time=nothing,
    total_elevation_gain=nothing,
    type=nothing,
    sport_type=nothing,
    workout_type=nothing,
)

- athlete::MetaAthlete
- name::String : The name of the activity
- distance::Float32 : The activity&#39;s distance, in meters
- moving_time::Int64 : The activity&#39;s moving time, in seconds
- elapsed_time::Int64 : The activity&#39;s elapsed time, in seconds
- total_elevation_gain::Float32 : The activity&#39;s total elevation gain.
- type::ActivityType
- sport_type::SportType
- workout_type::Int64 : The activity&#39;s workout type

StravaAPI.ClubAthlete — Type

ClubAthlete

ClubAthlete(;
    resource_state=nothing,
    firstname=nothing,
    lastname=nothing,
    member=nothing,
    admin=nothing,
    owner=nothing,
)

- resource_state::Int64 : Resource state, indicates level of detail. Possible values: 1 -&gt; \&quot;meta\&quot;, 2 -&gt; \&quot;summary\&quot;, 3 -&gt; \&quot;detail\&quot;
- firstname::String : The athlete&#39;s first name.
- lastname::String : The athlete&#39;s last initial.
- member::String : The athlete&#39;s member status.
- admin::Bool : Whether the athlete is a club admin.
- owner::Bool : Whether the athlete is club owner.

StravaAPI.Comment — Type

Comment

Comment(;
    id=nothing,
    activity_id=nothing,
    text=nothing,
    athlete=nothing,
    created_at=nothing,
)

- id::Int64 : The unique identifier of this comment
- activity_id::Int64 : The identifier of the activity this comment is related to
- text::String : The content of the comment
- athlete::SummaryAthlete
- created_at::ZonedDateTime : The time at which this comment was created.

StravaAPI.DetailedActivity — Type

DetailedActivity

DetailedActivity(;
    id=nothing,
    external_id=nothing,
    upload_id=nothing,
    athlete=nothing,
    name=nothing,
    distance=nothing,
    moving_time=nothing,
    elapsed_time=nothing,
    total_elevation_gain=nothing,
    elev_high=nothing,
    elev_low=nothing,
    type=nothing,
    sport_type=nothing,
    start_date=nothing,
    start_date_local=nothing,
    timezone=nothing,
    start_latlng=nothing,
    end_latlng=nothing,
    achievement_count=nothing,
    kudos_count=nothing,
    comment_count=nothing,
    athlete_count=nothing,
    photo_count=nothing,
    total_photo_count=nothing,
    map=nothing,
    trainer=nothing,
    commute=nothing,
    manual=nothing,
    private=nothing,
    flagged=nothing,
    workout_type=nothing,
    upload_id_str=nothing,
    average_speed=nothing,
    max_speed=nothing,
    has_kudoed=nothing,
    hide_from_home=nothing,
    gear_id=nothing,
    kilojoules=nothing,
    average_watts=nothing,
    device_watts=nothing,
    max_watts=nothing,
    weighted_average_watts=nothing,
    description=nothing,
    photos=nothing,
    gear=nothing,
    calories=nothing,
    segment_efforts=nothing,
    device_name=nothing,
    embed_token=nothing,
    splits_metric=nothing,
    splits_standard=nothing,
    laps=nothing,
    best_efforts=nothing,
)

- id::Int64 : The unique identifier of the activity
- external_id::String : The identifier provided at upload time
- upload_id::Int64 : The identifier of the upload that resulted in this activity
- athlete::MetaAthlete
- name::String : The name of the activity
- distance::Float32 : The activity&#39;s distance, in meters
- moving_time::Int64 : The activity&#39;s moving time, in seconds
- elapsed_time::Int64 : The activity&#39;s elapsed time, in seconds
- total_elevation_gain::Float32 : The activity&#39;s total elevation gain.
- elev_high::Float32 : The activity&#39;s highest elevation, in meters
- elev_low::Float32 : The activity&#39;s lowest elevation, in meters
- type::ActivityType
- sport_type::SportType
- start_date::ZonedDateTime : The time at which the activity was started.
- start_date_local::ZonedDateTime : The time at which the activity was started in the local timezone.
- timezone::String : The timezone of the activity
- start_latlng::Vector{Float32} : A pair of latitude/longitude coordinates, represented as an array of 2 floating point numbers.
- end_latlng::Vector{Float32} : A pair of latitude/longitude coordinates, represented as an array of 2 floating point numbers.
- achievement_count::Int64 : The number of achievements gained during this activity
- kudos_count::Int64 : The number of kudos given for this activity
- comment_count::Int64 : The number of comments for this activity
- athlete_count::Int64 : The number of athletes for taking part in a group activity
- photo_count::Int64 : The number of Instagram photos for this activity
- total_photo_count::Int64 : The number of Instagram and Strava photos for this activity
- map::PolylineMap
- trainer::Bool : Whether this activity was recorded on a training machine
- commute::Bool : Whether this activity is a commute
- manual::Bool : Whether this activity was created manually
- private::Bool : Whether this activity is private
- flagged::Bool : Whether this activity is flagged
- workout_type::Int64 : The activity&#39;s workout type
- upload_id_str::String : The unique identifier of the upload in string format
- average_speed::Float32 : The activity&#39;s average speed, in meters per second
- max_speed::Float32 : The activity&#39;s max speed, in meters per second
- has_kudoed::Bool : Whether the logged-in athlete has kudoed this activity
- hide_from_home::Bool : Whether the activity is muted
- gear_id::String : The id of the gear for the activity
- kilojoules::Float32 : The total work done in kilojoules during this activity. Rides only
- average_watts::Float32 : Average power output in watts during this activity. Rides only
- device_watts::Bool : Whether the watts are from a power meter, false if estimated
- max_watts::Int64 : Rides with power meter data only
- weighted_average_watts::Int64 : Similar to Normalized Power. Rides with power meter data only
- description::String : The description of the activity
- photos::PhotosSummary
- gear::SummaryGear
- calories::Float32 : The number of kilocalories consumed during this activity
- segment_efforts::Vector{DetailedSegmentEffort}
- device_name::String : The name of the device used to record the activity
- embed_token::String : The token used to embed a Strava activity
- splits_metric::Vector{Split} : The splits of this activity in metric units (for runs)
- splits_standard::Vector{Split} : The splits of this activity in imperial units (for runs)
- laps::Vector{Lap}
- best_efforts::Vector{DetailedSegmentEffort}

StravaAPI.DetailedAthlete — Type

DetailedAthlete

DetailedAthlete(;
    id=nothing,
    resource_state=nothing,
    firstname=nothing,
    lastname=nothing,
    profile_medium=nothing,
    profile=nothing,
    city=nothing,
    state=nothing,
    country=nothing,
    sex=nothing,
    premium=nothing,
    summit=nothing,
    created_at=nothing,
    updated_at=nothing,
    follower_count=nothing,
    friend_count=nothing,
    measurement_preference=nothing,
    ftp=nothing,
    weight=nothing,
    clubs=nothing,
    bikes=nothing,
    shoes=nothing,
)

- id::Int64 : The unique identifier of the athlete
- resource_state::Int64 : Resource state, indicates level of detail. Possible values: 1 -&gt; \&quot;meta\&quot;, 2 -&gt; \&quot;summary\&quot;, 3 -&gt; \&quot;detail\&quot;
- firstname::String : The athlete&#39;s first name.
- lastname::String : The athlete&#39;s last name.
- profile_medium::String : URL to a 62x62 pixel profile picture.
- profile::String : URL to a 124x124 pixel profile picture.
- city::String : The athlete&#39;s city.
- state::String : The athlete&#39;s state or geographical region.
- country::String : The athlete&#39;s country.
- sex::String : The athlete&#39;s sex.
- premium::Bool : Deprecated.  Use summit field instead. Whether the athlete has any Summit subscription.
- summit::Bool : Whether the athlete has any Summit subscription.
- created_at::ZonedDateTime : The time at which the athlete was created.
- updated_at::ZonedDateTime : The time at which the athlete was last updated.
- follower_count::Int64 : The athlete&#39;s follower count.
- friend_count::Int64 : The athlete&#39;s friend count.
- measurement_preference::String : The athlete&#39;s preferred unit system.
- ftp::Int64 : The athlete&#39;s FTP (Functional Threshold Power).
- weight::Float32 : The athlete&#39;s weight.
- clubs::Vector{SummaryClub} : The athlete&#39;s clubs.
- bikes::Vector{SummaryGear} : The athlete&#39;s bikes.
- shoes::Vector{SummaryGear} : The athlete&#39;s shoes.

StravaAPI.DetailedClub — Type

DetailedClub

DetailedClub(;
    id=nothing,
    resource_state=nothing,
    name=nothing,
    profile_medium=nothing,
    cover_photo=nothing,
    cover_photo_small=nothing,
    sport_type=nothing,
    activity_types=nothing,
    city=nothing,
    state=nothing,
    country=nothing,
    private=nothing,
    member_count=nothing,
    featured=nothing,
    verified=nothing,
    url=nothing,
    membership=nothing,
    admin=nothing,
    owner=nothing,
    following_count=nothing,
)

- id::Int64 : The club&#39;s unique identifier.
- resource_state::Int64 : Resource state, indicates level of detail. Possible values: 1 -&gt; \&quot;meta\&quot;, 2 -&gt; \&quot;summary\&quot;, 3 -&gt; \&quot;detail\&quot;
- name::String : The club&#39;s name.
- profile_medium::String : URL to a 60x60 pixel profile picture.
- cover_photo::String : URL to a ~1185x580 pixel cover photo.
- cover_photo_small::String : URL to a ~360x176  pixel cover photo.
- sport_type::String : Deprecated. Prefer to use activity_types.
- activity_types::Vector{ActivityType} : The activity types that count for a club. This takes precedence over sport_type.
- city::String : The club&#39;s city.
- state::String : The club&#39;s state or geographical region.
- country::String : The club&#39;s country.
- private::Bool : Whether the club is private.
- member_count::Int64 : The club&#39;s member count.
- featured::Bool : Whether the club is featured or not.
- verified::Bool : Whether the club is verified or not.
- url::String : The club&#39;s vanity URL.
- membership::String : The membership status of the logged-in athlete.
- admin::Bool : Whether the currently logged-in athlete is an administrator of this club.
- owner::Bool : Whether the currently logged-in athlete is the owner of this club.
- following_count::Int64 : The number of athletes in the club that the logged-in athlete follows.

StravaAPI.DetailedGear — Type

DetailedGear

DetailedGear(;
    id=nothing,
    resource_state=nothing,
    primary=nothing,
    name=nothing,
    distance=nothing,
    brand_name=nothing,
    model_name=nothing,
    frame_type=nothing,
    description=nothing,
)

- id::String : The gear&#39;s unique identifier.
- resource_state::Int64 : Resource state, indicates level of detail. Possible values: 2 -&gt; \&quot;summary\&quot;, 3 -&gt; \&quot;detail\&quot;
- primary::Bool : Whether this gear&#39;s is the owner&#39;s default one.
- name::String : The gear&#39;s name.
- distance::Float32 : The distance logged with this gear.
- brand_name::String : The gear&#39;s brand name.
- model_name::String : The gear&#39;s model name.
- frame_type::Int64 : The gear&#39;s frame type (bike only).
- description::String : The gear&#39;s description.

StravaAPI.DetailedSegment — Type

DetailedSegment

DetailedSegment(;
    id=nothing,
    name=nothing,
    activity_type=nothing,
    distance=nothing,
    average_grade=nothing,
    maximum_grade=nothing,
    elevation_high=nothing,
    elevation_low=nothing,
    start_latlng=nothing,
    end_latlng=nothing,
    climb_category=nothing,
    city=nothing,
    state=nothing,
    country=nothing,
    private=nothing,
    athlete_pr_effort=nothing,
    athlete_segment_stats=nothing,
    created_at=nothing,
    updated_at=nothing,
    total_elevation_gain=nothing,
    map=nothing,
    effort_count=nothing,
    athlete_count=nothing,
    hazardous=nothing,
    star_count=nothing,
)

- id::Int64 : The unique identifier of this segment
- name::String : The name of this segment
- activity_type::String
- distance::Float32 : The segment&#39;s distance, in meters
- average_grade::Float32 : The segment&#39;s average grade, in percents
- maximum_grade::Float32 : The segments&#39;s maximum grade, in percents
- elevation_high::Float32 : The segments&#39;s highest elevation, in meters
- elevation_low::Float32 : The segments&#39;s lowest elevation, in meters
- start_latlng::Vector{Float32} : A pair of latitude/longitude coordinates, represented as an array of 2 floating point numbers.
- end_latlng::Vector{Float32} : A pair of latitude/longitude coordinates, represented as an array of 2 floating point numbers.
- climb_category::Int64 : The category of the climb [0, 5]. Higher is harder ie. 5 is Hors catégorie, 0 is uncategorized in climb_category.
- city::String : The segments&#39;s city.
- state::String : The segments&#39;s state or geographical region.
- country::String : The segment&#39;s country.
- private::Bool : Whether this segment is private.
- athlete_pr_effort::SummaryPRSegmentEffort
- athlete_segment_stats::SummarySegmentEffort
- created_at::ZonedDateTime : The time at which the segment was created.
- updated_at::ZonedDateTime : The time at which the segment was last updated.
- total_elevation_gain::Float32 : The segment&#39;s total elevation gain.
- map::PolylineMap
- effort_count::Int64 : The total number of efforts for this segment
- athlete_count::Int64 : The number of unique athletes who have an effort for this segment
- hazardous::Bool : Whether this segment is considered hazardous
- star_count::Int64 : The number of stars for this segment

StravaAPI.DetailedSegmentEffort — Type

DetailedSegmentEffort

DetailedSegmentEffort(;
    id=nothing,
    activity_id=nothing,
    elapsed_time=nothing,
    start_date=nothing,
    start_date_local=nothing,
    distance=nothing,
    is_kom=nothing,
    name=nothing,
    activity=nothing,
    athlete=nothing,
    moving_time=nothing,
    start_index=nothing,
    end_index=nothing,
    average_cadence=nothing,
    average_watts=nothing,
    device_watts=nothing,
    average_heartrate=nothing,
    max_heartrate=nothing,
    segment=nothing,
    kom_rank=nothing,
    pr_rank=nothing,
    hidden=nothing,
)

- id::Int64 : The unique identifier of this effort
- activity_id::Int64 : The unique identifier of the activity related to this effort
- elapsed_time::Int64 : The effort&#39;s elapsed time
- start_date::ZonedDateTime : The time at which the effort was started.
- start_date_local::ZonedDateTime : The time at which the effort was started in the local timezone.
- distance::Float32 : The effort&#39;s distance in meters
- is_kom::Bool : Whether this effort is the current best on the leaderboard
- name::String : The name of the segment on which this effort was performed
- activity::MetaActivity
- athlete::MetaAthlete
- moving_time::Int64 : The effort&#39;s moving time
- start_index::Int64 : The start index of this effort in its activity&#39;s stream
- end_index::Int64 : The end index of this effort in its activity&#39;s stream
- average_cadence::Float32 : The effort&#39;s average cadence
- average_watts::Float32 : The average wattage of this effort
- device_watts::Bool : For riding efforts, whether the wattage was reported by a dedicated recording device
- average_heartrate::Float32 : The heart heart rate of the athlete during this effort
- max_heartrate::Float32 : The maximum heart rate of the athlete during this effort
- segment::SummarySegment
- kom_rank::Int64 : The rank of the effort on the global leaderboard if it belongs in the top 10 at the time of upload
- pr_rank::Int64 : The rank of the effort on the athlete&#39;s leaderboard if it belongs in the top 3 at the time of upload
- hidden::Bool : Whether this effort should be hidden when viewed within an activity

StravaAPI.DistanceStream — Type

DistanceStream

DistanceStream(;
    original_size=nothing,
    resolution=nothing,
    series_type=nothing,
    data=nothing,
)

- original_size::Int64 : The number of data points in this stream
- resolution::String : The level of detail (sampling) in which this stream was returned
- series_type::String : The base series used in the case the stream was downsampled
- data::Vector{Float32} : The sequence of distance values for this stream, in meters

StravaAPI.Error — Type

Error

Error(;
    code=nothing,
    field=nothing,
    resource=nothing,
)

- code::String : The code associated with this error.
- field::String : The specific field or aspect of the resource associated with this error.
- resource::String : The type of resource associated with this error.

StravaAPI.ExplorerResponse — Type

ExplorerResponse

ExplorerResponse(;
    segments=nothing,
)

- segments::Vector{ExplorerSegment} : The set of segments matching an explorer request

StravaAPI.ExplorerSegment — Type

ExplorerSegment

ExplorerSegment(;
    id=nothing,
    name=nothing,
    climb_category=nothing,
    climb_category_desc=nothing,
    avg_grade=nothing,
    start_latlng=nothing,
    end_latlng=nothing,
    elev_difference=nothing,
    distance=nothing,
    points=nothing,
)

- id::Int64 : The unique identifier of this segment
- name::String : The name of this segment
- climb_category::Int64 : The category of the climb [0, 5]. Higher is harder ie. 5 is Hors catégorie, 0 is uncategorized in climb_category. If climb_category &#x3D; 5, climb_category_desc &#x3D; HC. If climb_category &#x3D; 2, climb_category_desc &#x3D; 3.
- climb_category_desc::String : The description for the category of the climb
- avg_grade::Float32 : The segment&#39;s average grade, in percents
- start_latlng::Vector{Float32} : A pair of latitude/longitude coordinates, represented as an array of 2 floating point numbers.
- end_latlng::Vector{Float32} : A pair of latitude/longitude coordinates, represented as an array of 2 floating point numbers.
- elev_difference::Float32 : The segments&#39;s evelation difference, in meters
- distance::Float32 : The segment&#39;s distance, in meters
- points::String : The polyline of the segment

StravaAPI.Fault — Type

Fault Encapsulates the errors that may be returned from the API.

Fault(;
    errors=nothing,
    message=nothing,
)

- errors::Vector{Error} : The set of specific errors associated with this fault, if any.
- message::String : The message of the fault.

StravaAPI.HeartRateZoneRanges — Type

HeartRateZoneRanges

HeartRateZoneRanges(;
    custom_zones=nothing,
    zones=nothing,
)

- custom_zones::Bool : Whether the athlete has set their own custom heart rate zones
- zones::Vector{ZoneRange}

StravaAPI.HeartrateStream — Type

HeartrateStream

HeartrateStream(;
    original_size=nothing,
    resolution=nothing,
    series_type=nothing,
    data=nothing,
)

- original_size::Int64 : The number of data points in this stream
- resolution::String : The level of detail (sampling) in which this stream was returned
- series_type::String : The base series used in the case the stream was downsampled
- data::Vector{Int64} : The sequence of heart rate values for this stream, in beats per minute

StravaAPI.Lap — Type

Lap

Lap(;
    id=nothing,
    activity=nothing,
    athlete=nothing,
    average_cadence=nothing,
    average_speed=nothing,
    distance=nothing,
    elapsed_time=nothing,
    start_index=nothing,
    end_index=nothing,
    lap_index=nothing,
    max_speed=nothing,
    moving_time=nothing,
    name=nothing,
    pace_zone=nothing,
    split=nothing,
    start_date=nothing,
    start_date_local=nothing,
    total_elevation_gain=nothing,
)

- id::Int64 : The unique identifier of this lap
- activity::MetaActivity
- athlete::MetaAthlete
- average_cadence::Float32 : The lap&#39;s average cadence
- average_speed::Float32 : The lap&#39;s average speed
- distance::Float32 : The lap&#39;s distance, in meters
- elapsed_time::Int64 : The lap&#39;s elapsed time, in seconds
- start_index::Int64 : The start index of this effort in its activity&#39;s stream
- end_index::Int64 : The end index of this effort in its activity&#39;s stream
- lap_index::Int64 : The index of this lap in the activity it belongs to
- max_speed::Float32 : The maximum speed of this lat, in meters per second
- moving_time::Int64 : The lap&#39;s moving time, in seconds
- name::String : The name of the lap
- pace_zone::Int64 : The athlete&#39;s pace zone during this lap
- split::Int64
- start_date::ZonedDateTime : The time at which the lap was started.
- start_date_local::ZonedDateTime : The time at which the lap was started in the local timezone.
- total_elevation_gain::Float32 : The elevation gain of this lap, in meters

StravaAPI.LatLngStream — Type

LatLngStream

LatLngStream(;
    original_size=nothing,
    resolution=nothing,
    series_type=nothing,
    data=nothing,
)

- original_size::Int64 : The number of data points in this stream
- resolution::String : The level of detail (sampling) in which this stream was returned
- series_type::String : The base series used in the case the stream was downsampled
- data::Vector{Vector} : The sequence of lat/long values for this stream

StravaAPI.MetaActivity — Type

MetaActivity

MetaActivity(;
    id=nothing,
)

- id::Int64 : The unique identifier of the activity

StravaAPI.MetaAthlete — Type

MetaAthlete

MetaAthlete(;
    id=nothing,
)

- id::Int64 : The unique identifier of the athlete

StravaAPI.MetaClub — Type

MetaClub

MetaClub(;
    id=nothing,
    resource_state=nothing,
    name=nothing,
)

- id::Int64 : The club&#39;s unique identifier.
- resource_state::Int64 : Resource state, indicates level of detail. Possible values: 1 -&gt; \&quot;meta\&quot;, 2 -&gt; \&quot;summary\&quot;, 3 -&gt; \&quot;detail\&quot;
- name::String : The club&#39;s name.

StravaAPI.MovingStream — Type

MovingStream

MovingStream(;
    original_size=nothing,
    resolution=nothing,
    series_type=nothing,
    data=nothing,
)

- original_size::Int64 : The number of data points in this stream
- resolution::String : The level of detail (sampling) in which this stream was returned
- series_type::String : The base series used in the case the stream was downsampled
- data::Vector{Bool} : The sequence of moving values for this stream, as boolean values

StravaAPI.PhotosSummary — Type

PhotosSummary

PhotosSummary(;
    count=nothing,
    primary=nothing,
)

- count::Int64 : The number of photos
- primary::PhotosSummaryPrimary

StravaAPI.PhotosSummaryPrimary — Type

PhotosSummary_primary

PhotosSummaryPrimary(;
    id=nothing,
    source=nothing,
    unique_id=nothing,
    urls=nothing,
)

- id::Int64
- source::Int64
- unique_id::String
- urls::Dict{String, String}

StravaAPI.PolylineMap — Type

PolylineMap

PolylineMap(;
    id=nothing,
    polyline=nothing,
    summary_polyline=nothing,
)

- id::String : The identifier of the map
- polyline::String : The polyline of the map, only returned on detailed representation of an object
- summary_polyline::String : The summary polyline of the map

StravaAPI.PowerStream — Type

PowerStream

PowerStream(;
    original_size=nothing,
    resolution=nothing,
    series_type=nothing,
    data=nothing,
)

- original_size::Int64 : The number of data points in this stream
- resolution::String : The level of detail (sampling) in which this stream was returned
- series_type::String : The base series used in the case the stream was downsampled
- data::Vector{Int64} : The sequence of power values for this stream, in watts

StravaAPI.PowerZoneRanges — Type

PowerZoneRanges

PowerZoneRanges(;
    zones=nothing,
)

- zones::Vector{ZoneRange}

StravaAPI.Route — Type

Route

Route(;
    athlete=nothing,
    description=nothing,
    distance=nothing,
    elevation_gain=nothing,
    id=nothing,
    id_str=nothing,
    map=nothing,
    name=nothing,
    private=nothing,
    starred=nothing,
    timestamp=nothing,
    type=nothing,
    sub_type=nothing,
    created_at=nothing,
    updated_at=nothing,
    estimated_moving_time=nothing,
    segments=nothing,
    waypoints=nothing,
)

- athlete::SummaryAthlete
- description::String : The description of the route
- distance::Float32 : The route&#39;s distance, in meters
- elevation_gain::Float32 : The route&#39;s elevation gain.
- id::Int64 : The unique identifier of this route
- id_str::String : The unique identifier of the route in string format
- map::PolylineMap
- name::String : The name of this route
- private::Bool : Whether this route is private
- starred::Bool : Whether this route is starred by the logged-in athlete
- timestamp::Int64 : An epoch timestamp of when the route was created
- type::Int64 : This route&#39;s type (1 for ride, 2 for runs)
- sub_type::Int64 : This route&#39;s sub-type (1 for road, 2 for mountain bike, 3 for cross, 4 for trail, 5 for mixed)
- created_at::ZonedDateTime : The time at which the route was created
- updated_at::ZonedDateTime : The time at which the route was last updated
- estimated_moving_time::Int64 : Estimated time in seconds for the authenticated athlete to complete route
- segments::Vector{SummarySegment} : The segments traversed by this route
- waypoints::Vector{Waypoint} : The custom waypoints along this route

StravaAPI.SmoothGradeStream — Type

SmoothGradeStream

SmoothGradeStream(;
    original_size=nothing,
    resolution=nothing,
    series_type=nothing,
    data=nothing,
)

- original_size::Int64 : The number of data points in this stream
- resolution::String : The level of detail (sampling) in which this stream was returned
- series_type::String : The base series used in the case the stream was downsampled
- data::Vector{Float32} : The sequence of grade values for this stream, as percents of a grade

StravaAPI.SmoothVelocityStream — Type

SmoothVelocityStream

SmoothVelocityStream(;
    original_size=nothing,
    resolution=nothing,
    series_type=nothing,
    data=nothing,
)

- original_size::Int64 : The number of data points in this stream
- resolution::String : The level of detail (sampling) in which this stream was returned
- series_type::String : The base series used in the case the stream was downsampled
- data::Vector{Float32} : The sequence of velocity values for this stream, in meters per second

StravaAPI.Split — Type

Split

Split(;
    average_speed=nothing,
    distance=nothing,
    elapsed_time=nothing,
    elevation_difference=nothing,
    pace_zone=nothing,
    moving_time=nothing,
    split=nothing,
)

- average_speed::Float32 : The average speed of this split, in meters per second
- distance::Float32 : The distance of this split, in meters
- elapsed_time::Int64 : The elapsed time of this split, in seconds
- elevation_difference::Float32 : The elevation difference of this split, in meters
- pace_zone::Int64 : The pacing zone of this split
- moving_time::Int64 : The moving time of this split, in seconds
- split::Int64 : N/A

StravaAPI.StreamSet — Type

StreamSet

StreamSet(;
    time=nothing,
    distance=nothing,
    latlng=nothing,
    altitude=nothing,
    velocity_smooth=nothing,
    heartrate=nothing,
    cadence=nothing,
    watts=nothing,
    temp=nothing,
    moving=nothing,
    grade_smooth=nothing,
)

- time::TimeStream
- distance::DistanceStream
- latlng::LatLngStream
- altitude::AltitudeStream
- velocity_smooth::SmoothVelocityStream
- heartrate::HeartrateStream
- cadence::CadenceStream
- watts::PowerStream
- temp::TemperatureStream
- moving::MovingStream
- grade_smooth::SmoothGradeStream

StravaAPI.SummaryActivity — Type

SummaryActivity

SummaryActivity(;
    id=nothing,
    external_id=nothing,
    upload_id=nothing,
    athlete=nothing,
    name=nothing,
    distance=nothing,
    moving_time=nothing,
    elapsed_time=nothing,
    total_elevation_gain=nothing,
    elev_high=nothing,
    elev_low=nothing,
    type=nothing,
    sport_type=nothing,
    start_date=nothing,
    start_date_local=nothing,
    timezone=nothing,
    start_latlng=nothing,
    end_latlng=nothing,
    achievement_count=nothing,
    kudos_count=nothing,
    comment_count=nothing,
    athlete_count=nothing,
    photo_count=nothing,
    total_photo_count=nothing,
    map=nothing,
    trainer=nothing,
    commute=nothing,
    manual=nothing,
    private=nothing,
    flagged=nothing,
    workout_type=nothing,
    upload_id_str=nothing,
    average_speed=nothing,
    max_speed=nothing,
    has_kudoed=nothing,
    hide_from_home=nothing,
    gear_id=nothing,
    kilojoules=nothing,
    average_watts=nothing,
    device_watts=nothing,
    max_watts=nothing,
    weighted_average_watts=nothing,
)

- id::Int64 : The unique identifier of the activity
- external_id::String : The identifier provided at upload time
- upload_id::Int64 : The identifier of the upload that resulted in this activity
- athlete::MetaAthlete
- name::String : The name of the activity
- distance::Float32 : The activity&#39;s distance, in meters
- moving_time::Int64 : The activity&#39;s moving time, in seconds
- elapsed_time::Int64 : The activity&#39;s elapsed time, in seconds
- total_elevation_gain::Float32 : The activity&#39;s total elevation gain.
- elev_high::Float32 : The activity&#39;s highest elevation, in meters
- elev_low::Float32 : The activity&#39;s lowest elevation, in meters
- type::ActivityType
- sport_type::SportType
- start_date::ZonedDateTime : The time at which the activity was started.
- start_date_local::ZonedDateTime : The time at which the activity was started in the local timezone.
- timezone::String : The timezone of the activity
- start_latlng::Vector{Float32} : A pair of latitude/longitude coordinates, represented as an array of 2 floating point numbers.
- end_latlng::Vector{Float32} : A pair of latitude/longitude coordinates, represented as an array of 2 floating point numbers.
- achievement_count::Int64 : The number of achievements gained during this activity
- kudos_count::Int64 : The number of kudos given for this activity
- comment_count::Int64 : The number of comments for this activity
- athlete_count::Int64 : The number of athletes for taking part in a group activity
- photo_count::Int64 : The number of Instagram photos for this activity
- total_photo_count::Int64 : The number of Instagram and Strava photos for this activity
- map::PolylineMap
- trainer::Bool : Whether this activity was recorded on a training machine
- commute::Bool : Whether this activity is a commute
- manual::Bool : Whether this activity was created manually
- private::Bool : Whether this activity is private
- flagged::Bool : Whether this activity is flagged
- workout_type::Int64 : The activity&#39;s workout type
- upload_id_str::String : The unique identifier of the upload in string format
- average_speed::Float32 : The activity&#39;s average speed, in meters per second
- max_speed::Float32 : The activity&#39;s max speed, in meters per second
- has_kudoed::Bool : Whether the logged-in athlete has kudoed this activity
- hide_from_home::Bool : Whether the activity is muted
- gear_id::String : The id of the gear for the activity
- kilojoules::Float32 : The total work done in kilojoules during this activity. Rides only
- average_watts::Float32 : Average power output in watts during this activity. Rides only
- device_watts::Bool : Whether the watts are from a power meter, false if estimated
- max_watts::Int64 : Rides with power meter data only
- weighted_average_watts::Int64 : Similar to Normalized Power. Rides with power meter data only

StravaAPI.SummaryAthlete — Type

SummaryAthlete

SummaryAthlete(;
    id=nothing,
    resource_state=nothing,
    firstname=nothing,
    lastname=nothing,
    profile_medium=nothing,
    profile=nothing,
    city=nothing,
    state=nothing,
    country=nothing,
    sex=nothing,
    premium=nothing,
    summit=nothing,
    created_at=nothing,
    updated_at=nothing,
)

- id::Int64 : The unique identifier of the athlete
- resource_state::Int64 : Resource state, indicates level of detail. Possible values: 1 -&gt; \&quot;meta\&quot;, 2 -&gt; \&quot;summary\&quot;, 3 -&gt; \&quot;detail\&quot;
- firstname::String : The athlete&#39;s first name.
- lastname::String : The athlete&#39;s last name.
- profile_medium::String : URL to a 62x62 pixel profile picture.
- profile::String : URL to a 124x124 pixel profile picture.
- city::String : The athlete&#39;s city.
- state::String : The athlete&#39;s state or geographical region.
- country::String : The athlete&#39;s country.
- sex::String : The athlete&#39;s sex.
- premium::Bool : Deprecated.  Use summit field instead. Whether the athlete has any Summit subscription.
- summit::Bool : Whether the athlete has any Summit subscription.
- created_at::ZonedDateTime : The time at which the athlete was created.
- updated_at::ZonedDateTime : The time at which the athlete was last updated.

StravaAPI.SummaryClub — Type

SummaryClub

SummaryClub(;
    id=nothing,
    resource_state=nothing,
    name=nothing,
    profile_medium=nothing,
    cover_photo=nothing,
    cover_photo_small=nothing,
    sport_type=nothing,
    activity_types=nothing,
    city=nothing,
    state=nothing,
    country=nothing,
    private=nothing,
    member_count=nothing,
    featured=nothing,
    verified=nothing,
    url=nothing,
)

- id::Int64 : The club&#39;s unique identifier.
- resource_state::Int64 : Resource state, indicates level of detail. Possible values: 1 -&gt; \&quot;meta\&quot;, 2 -&gt; \&quot;summary\&quot;, 3 -&gt; \&quot;detail\&quot;
- name::String : The club&#39;s name.
- profile_medium::String : URL to a 60x60 pixel profile picture.
- cover_photo::String : URL to a ~1185x580 pixel cover photo.
- cover_photo_small::String : URL to a ~360x176  pixel cover photo.
- sport_type::String : Deprecated. Prefer to use activity_types.
- activity_types::Vector{ActivityType} : The activity types that count for a club. This takes precedence over sport_type.
- city::String : The club&#39;s city.
- state::String : The club&#39;s state or geographical region.
- country::String : The club&#39;s country.
- private::Bool : Whether the club is private.
- member_count::Int64 : The club&#39;s member count.
- featured::Bool : Whether the club is featured or not.
- verified::Bool : Whether the club is verified or not.
- url::String : The club&#39;s vanity URL.

StravaAPI.SummaryGear — Type

SummaryGear

SummaryGear(;
    id=nothing,
    resource_state=nothing,
    primary=nothing,
    name=nothing,
    distance=nothing,
)

- id::String : The gear&#39;s unique identifier.
- resource_state::Int64 : Resource state, indicates level of detail. Possible values: 2 -&gt; \&quot;summary\&quot;, 3 -&gt; \&quot;detail\&quot;
- primary::Bool : Whether this gear&#39;s is the owner&#39;s default one.
- name::String : The gear&#39;s name.
- distance::Float32 : The distance logged with this gear.

StravaAPI.SummaryPRSegmentEffort — Type

SummaryPRSegmentEffort

SummaryPRSegmentEffort(;
    pr_activity_id=nothing,
    pr_elapsed_time=nothing,
    pr_date=nothing,
    effort_count=nothing,
)

- pr_activity_id::Int64 : The unique identifier of the activity related to the PR effort.
- pr_elapsed_time::Int64 : The elapsed time ot the PR effort.
- pr_date::ZonedDateTime : The time at which the PR effort was started.
- effort_count::Int64 : Number of efforts by the authenticated athlete on this segment.

StravaAPI.SummarySegment — Type

SummarySegment

SummarySegment(;
    id=nothing,
    name=nothing,
    activity_type=nothing,
    distance=nothing,
    average_grade=nothing,
    maximum_grade=nothing,
    elevation_high=nothing,
    elevation_low=nothing,
    start_latlng=nothing,
    end_latlng=nothing,
    climb_category=nothing,
    city=nothing,
    state=nothing,
    country=nothing,
    private=nothing,
    athlete_pr_effort=nothing,
    athlete_segment_stats=nothing,
)

- id::Int64 : The unique identifier of this segment
- name::String : The name of this segment
- activity_type::String
- distance::Float32 : The segment&#39;s distance, in meters
- average_grade::Float32 : The segment&#39;s average grade, in percents
- maximum_grade::Float32 : The segments&#39;s maximum grade, in percents
- elevation_high::Float32 : The segments&#39;s highest elevation, in meters
- elevation_low::Float32 : The segments&#39;s lowest elevation, in meters
- start_latlng::Vector{Float32} : A pair of latitude/longitude coordinates, represented as an array of 2 floating point numbers.
- end_latlng::Vector{Float32} : A pair of latitude/longitude coordinates, represented as an array of 2 floating point numbers.
- climb_category::Int64 : The category of the climb [0, 5]. Higher is harder ie. 5 is Hors catégorie, 0 is uncategorized in climb_category.
- city::String : The segments&#39;s city.
- state::String : The segments&#39;s state or geographical region.
- country::String : The segment&#39;s country.
- private::Bool : Whether this segment is private.
- athlete_pr_effort::SummaryPRSegmentEffort
- athlete_segment_stats::SummarySegmentEffort

StravaAPI.SummarySegmentEffort — Type

SummarySegmentEffort

SummarySegmentEffort(;
    id=nothing,
    activity_id=nothing,
    elapsed_time=nothing,
    start_date=nothing,
    start_date_local=nothing,
    distance=nothing,
    is_kom=nothing,
)

- id::Int64 : The unique identifier of this effort
- activity_id::Int64 : The unique identifier of the activity related to this effort
- elapsed_time::Int64 : The effort&#39;s elapsed time
- start_date::ZonedDateTime : The time at which the effort was started.
- start_date_local::ZonedDateTime : The time at which the effort was started in the local timezone.
- distance::Float32 : The effort&#39;s distance in meters
- is_kom::Bool : Whether this effort is the current best on the leaderboard

StravaAPI.TemperatureStream — Type

TemperatureStream

TemperatureStream(;
    original_size=nothing,
    resolution=nothing,
    series_type=nothing,
    data=nothing,
)

- original_size::Int64 : The number of data points in this stream
- resolution::String : The level of detail (sampling) in which this stream was returned
- series_type::String : The base series used in the case the stream was downsampled
- data::Vector{Int64} : The sequence of temperature values for this stream, in celsius degrees

StravaAPI.TimeStream — Type

TimeStream

TimeStream(;
    original_size=nothing,
    resolution=nothing,
    series_type=nothing,
    data=nothing,
)

- original_size::Int64 : The number of data points in this stream
- resolution::String : The level of detail (sampling) in which this stream was returned
- series_type::String : The base series used in the case the stream was downsampled
- data::Vector{Int64} : The sequence of time values for this stream, in seconds

StravaAPI.TimedZoneRange — Type

TimedZoneRange A union type representing the time spent in a given zone.

TimedZoneRange(;
    min=nothing,
    max=nothing,
    time=nothing,
)

- min::Int64 : The minimum value in the range.
- max::Int64 : The maximum value in the range.
- time::Int64 : The number of seconds spent in this zone

StravaAPI.UpdatableActivity — Type

UpdatableActivity

UpdatableActivity(;
    commute=nothing,
    trainer=nothing,
    hide_from_home=nothing,
    description=nothing,
    name=nothing,
    type=nothing,
    sport_type=nothing,
    gear_id=nothing,
)

- commute::Bool : Whether this activity is a commute
- trainer::Bool : Whether this activity was recorded on a training machine
- hide_from_home::Bool : Whether this activity is muted
- description::String : The description of the activity
- name::String : The name of the activity
- type::ActivityType
- sport_type::SportType
- gear_id::String : Identifier for the gear associated with the activity. ‘none’ clears gear from activity

StravaAPI.Upload — Type

Upload

Upload(;
    id=nothing,
    id_str=nothing,
    external_id=nothing,
    error=nothing,
    status=nothing,
    activity_id=nothing,
)

- id::Int64 : The unique identifier of the upload
- id_str::String : The unique identifier of the upload in string format
- external_id::String : The external identifier of the upload
- error::String : The error associated with this upload
- status::String : The status of this upload
- activity_id::Int64 : The identifier of the activity this upload resulted into

StravaAPI.Waypoint — Type

Waypoint

Waypoint(;
    latlng=nothing,
    target_latlng=nothing,
    categories=nothing,
    title=nothing,
    description=nothing,
    distance_into_route=nothing,
)

- latlng::Vector{Float32} : A pair of latitude/longitude coordinates, represented as an array of 2 floating point numbers.
- target_latlng::Vector{Float32} : A pair of latitude/longitude coordinates, represented as an array of 2 floating point numbers.
- categories::Vector{String} : Categories that the waypoint belongs to
- title::String : A title for the waypoint
- description::String : A description of the waypoint (optional)
- distance_into_route::Int64 : The number meters along the route that the waypoint is located

StravaAPI.ZoneRange — Type

ZoneRange

ZoneRange(;
    min=nothing,
    max=nothing,
)

- min::Int64 : The minimum value in the range.
- max::Int64 : The maximum value in the range.

StravaAPI.Zones — Type

Zones

Zones(;
    heart_rate=nothing,
    power=nothing,
)

- heart_rate::HeartRateZoneRanges
- power::PowerZoneRanges

StravaAPI.basepath — Method

The default API base path for APIs in ActivitiesApi. This can be used to construct the OpenAPI.Clients.Client instance.

StravaAPI.basepath — Method

The default API base path for APIs in AthletesApi. This can be used to construct the OpenAPI.Clients.Client instance.

StravaAPI.basepath — Method

The default API base path for APIs in ClubsApi. This can be used to construct the OpenAPI.Clients.Client instance.

StravaAPI.basepath — Method

The default API base path for APIs in GearsApi. This can be used to construct the OpenAPI.Clients.Client instance.

StravaAPI.basepath — Method

The default API base path for APIs in RoutesApi. This can be used to construct the OpenAPI.Clients.Client instance.

StravaAPI.basepath — Method

The default API base path for APIs in SegmentEffortsApi. This can be used to construct the OpenAPI.Clients.Client instance.

StravaAPI.basepath — Method

The default API base path for APIs in SegmentsApi. This can be used to construct the OpenAPI.Clients.Client instance.

StravaAPI.basepath — Method

The default API base path for APIs in StreamsApi. This can be used to construct the OpenAPI.Clients.Client instance.

StravaAPI.basepath — Method

The default API base path for APIs in UploadsApi. This can be used to construct the OpenAPI.Clients.Client instance.

StravaAPI.create_activity — Method

Create an Activity

Creates a manual activity for an athlete, requires activity:write scope.

Params:

name::String (required)

sport_type::String (required)

startdatelocal::ZonedDateTime (required)

elapsed_time::Int64 (required)

type::String

description::String

distance::Float32

trainer::Int64

commute::Int64

Return: DetailedActivity, OpenAPI.Clients.ApiResponse

StravaAPI.create_upload — Method

Upload Activity

Uploads a new data file to create an activity from. Requires activity:write scope.

Params:

file::String

name::String

description::String

trainer::String

commute::String

data_type::String

external_id::String

Return: Upload, OpenAPI.Clients.ApiResponse

StravaAPI.explore_segments — Method

Explore segments

Returns the top 10 segments matching a specified query.

Params:

bounds::Vector{Float32} (required)

activity_type::String

min_cat::Int64

max_cat::Int64

Return: ExplorerResponse, OpenAPI.Clients.ApiResponse

StravaAPI.get_activity_by_id — Method

Get Activity

Returns the given activity that is owned by the authenticated athlete. Requires activity:read for Everyone and Followers activities. Requires activity:read_all for Only Me activities.

Params:

id::Int64 (required)

includeallefforts::Bool

Return: DetailedActivity, OpenAPI.Clients.ApiResponse

StravaAPI.get_activity_streams — Method

Get Activity Streams

Returns the given activity's streams. Requires activity:read scope. Requires activity:read_all scope for Only Me activities.

Params:

id::Int64 (required)

keys::Vector{String} (required)

keybytype::Bool (required)

Return: StreamSet, OpenAPI.Clients.ApiResponse

StravaAPI.get_club_activities_by_id — Method

List Club Activities

Retrieve recent activities from members of a specific club. The authenticated athlete must belong to the requested club in order to hit this endpoint. Pagination is supported. Athlete profile visibility is respected for all activities.

Params:

id::Int64 (required)

page::Int64

per_page::Int64

Return: Vector{ClubActivity}, OpenAPI.Clients.ApiResponse

StravaAPI.get_club_admins_by_id — Method

List Club Administrators

Returns a list of the administrators of a given club.

Params:

id::Int64 (required)

page::Int64

per_page::Int64

Return: Vector{SummaryAthlete}, OpenAPI.Clients.ApiResponse

StravaAPI.get_club_by_id — Method

Get Club

Returns a given club using its identifier.

Params:

id::Int64 (required)

Return: DetailedClub, OpenAPI.Clients.ApiResponse

StravaAPI.get_club_members_by_id — Method

List Club Members

Returns a list of the athletes who are members of a given club.

Params:

id::Int64 (required)

page::Int64

per_page::Int64

Return: Vector{ClubAthlete}, OpenAPI.Clients.ApiResponse

StravaAPI.get_comments_by_activity_id — Method

List Activity Comments

Returns the comments on the given activity. Requires activity:read for Everyone and Followers activities. Requires activity:read_all for Only Me activities.

Params:

id::Int64 (required)

page::Int64

per_page::Int64

page_size::Int64

after_cursor::String

Return: Vector{Comment}, OpenAPI.Clients.ApiResponse

StravaAPI.get_efforts_by_segment_id — Method

List Segment Efforts

Returns a set of the authenticated athlete's segment efforts for a given segment. Requires subscription.

Params:

segment_id::Int64 (required)

startdatelocal::ZonedDateTime

enddatelocal::ZonedDateTime

per_page::Int64

Return: Vector{DetailedSegmentEffort}, OpenAPI.Clients.ApiResponse

StravaAPI.get_gear_by_id — Method

Get Equipment

Returns an equipment using its identifier.

Params:

id::String (required)

Return: DetailedGear, OpenAPI.Clients.ApiResponse

StravaAPI.get_kudoers_by_activity_id — Method

List Activity Kudoers

Returns the athletes who kudoed an activity identified by an identifier. Requires activity:read for Everyone and Followers activities. Requires activity:read_all for Only Me activities.

Params:

id::Int64 (required)

page::Int64

per_page::Int64

Return: Vector{SummaryAthlete}, OpenAPI.Clients.ApiResponse

StravaAPI.get_laps_by_activity_id — Method

List Activity Laps

Returns the laps of an activity identified by an identifier. Requires activity:read for Everyone and Followers activities. Requires activity:read_all for Only Me activities.

Params:

id::Int64 (required)

Return: Vector{Lap}, OpenAPI.Clients.ApiResponse

StravaAPI.get_logged_in_athlete — Method

Get Authenticated Athlete

Returns the currently authenticated athlete. Tokens with profile:read_all scope will receive a detailed athlete representation; all others will receive a summary representation.

Params:

Return: DetailedAthlete, OpenAPI.Clients.ApiResponse

StravaAPI.get_logged_in_athlete_activities — Method

List Athlete Activities

Returns the activities of an athlete for a specific identifier. Requires activity:read. Only Me activities will be filtered out unless requested by a token with activity:read_all.

Params:

before::Int64

after::Int64

page::Int64

per_page::Int64

Return: Vector{SummaryActivity}, OpenAPI.Clients.ApiResponse

StravaAPI.get_logged_in_athlete_clubs — Method

List Athlete Clubs

Returns a list of the clubs whose membership includes the authenticated athlete.

Params:

page::Int64

per_page::Int64

Return: Vector{SummaryClub}, OpenAPI.Clients.ApiResponse

StravaAPI.get_logged_in_athlete_starred_segments — Method

List Starred Segments

List of the authenticated athlete's starred segments. Private segments are filtered out unless requested by a token with read_all scope.

Params:

page::Int64

per_page::Int64

Return: Vector{SummarySegment}, OpenAPI.Clients.ApiResponse

StravaAPI.get_logged_in_athlete_zones — Method

Get Zones

Returns the the authenticated athlete's heart rate and power zones. Requires profile:read_all.

Params:

Return: Zones, OpenAPI.Clients.ApiResponse

StravaAPI.get_route_as_g_p_x — Method

Export Route GPX

Returns a GPX file of the route. Requires read_all scope for private routes.

Params:

id::Int64 (required)

Return: Nothing, OpenAPI.Clients.ApiResponse

StravaAPI.get_route_as_t_c_x — Method

Export Route TCX

Returns a TCX file of the route. Requires read_all scope for private routes.

Params:

id::Int64 (required)

Return: Nothing, OpenAPI.Clients.ApiResponse

StravaAPI.get_route_by_id — Method

Get Route

Returns a route using its identifier. Requires read_all scope for private routes.

Params:

id::Int64 (required)

Return: Route, OpenAPI.Clients.ApiResponse

StravaAPI.get_route_streams — Method

Get Route Streams

Returns the given route's streams. Requires read_all scope for private routes.

Params:

id::Int64 (required)

Return: StreamSet, OpenAPI.Clients.ApiResponse

StravaAPI.get_routes_by_athlete_id — Method

List Athlete Routes

Returns a list of the routes created by the authenticated athlete. Private routes are filtered out unless requested by a token with read_all scope.

Params:

page::Int64

per_page::Int64

Return: Vector{Route}, OpenAPI.Clients.ApiResponse

StravaAPI.get_segment_by_id — Method

Get Segment

Returns the specified segment. read_all scope required in order to retrieve athlete-specific segment information, or to retrieve private segments.

Params:

id::Int64 (required)

Return: DetailedSegment, OpenAPI.Clients.ApiResponse

StravaAPI.get_segment_effort_by_id — Method

Get Segment Effort

Returns a segment effort from an activity that is owned by the authenticated athlete. Requires subscription.

Params:

id::Int64 (required)

Return: DetailedSegmentEffort, OpenAPI.Clients.ApiResponse

StravaAPI.get_segment_effort_streams — Method

Get Segment Effort Streams

Returns a set of streams for a segment effort completed by the authenticated athlete. Requires read_all scope.

Params:

id::Int64 (required)

keys::Vector{String} (required)

keybytype::Bool (required)

Return: StreamSet, OpenAPI.Clients.ApiResponse

StravaAPI.get_segment_streams — Method

Get Segment Streams

Returns the given segment's streams. Requires read_all scope for private segments.

Params:

id::Int64 (required)

keys::Vector{String} (required)

keybytype::Bool (required)

Return: StreamSet, OpenAPI.Clients.ApiResponse

StravaAPI.get_stats — Method

Get Athlete Stats

Returns the activity stats of an athlete. Only includes data from activities set to Everyone visibilty.

Params:

id::Int64 (required)

Return: ActivityStats, OpenAPI.Clients.ApiResponse

StravaAPI.get_upload_by_id — Method

Get Upload

Returns an upload for a given identifier. Requires activity:write scope.

Params:

upload_id::Int64 (required)

Return: Upload, OpenAPI.Clients.ApiResponse

StravaAPI.get_zones_by_activity_id — Method

Get Activity Zones

Summit Feature. Returns the zones of a given activity. Requires activity:read for Everyone and Followers activities. Requires activity:read_all for Only Me activities.

Params:

id::Int64 (required)

Return: Vector{ActivityZone}, OpenAPI.Clients.ApiResponse

StravaAPI.star_segment — Method

Star Segment

Stars/Unstars the given segment for the authenticated athlete. Requires profile:write scope.

Params:

id::Int64 (required)

starred::Bool (required)

Return: DetailedSegment, OpenAPI.Clients.ApiResponse

StravaAPI.update_activity_by_id — Method

Update Activity

Updates the given activity that is owned by the authenticated athlete. Requires activity:write. Also requires activity:read_all in order to update Only Me activities

Params:

id::Int64 (required)

body::UpdatableActivity

Return: DetailedActivity, OpenAPI.Clients.ApiResponse

StravaAPI.update_logged_in_athlete — Method

Update Athlete

Update the currently authenticated athlete. Requires profile:write scope.

Params:

weight::Float32 (required)

Return: DetailedAthlete, OpenAPI.Clients.ApiResponse